pyproject.toml
[project]
name = "libvirt-aws"
description = "Partial AWS API emulation powered by libvirt"
dynamic = ["version"]
authors = [{ email = "elvis@edgedb.com", name = "Elvis Pranskevichus" }]
dependencies = [
    "aiohttp~=3.8.1",
    "click~=8.1.3",
    "dicttoxml~=1.7.16",
    "libvirt-python>=6.0.0",
    "xmltodict~=0.13.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = ["black~=22.3.0", "flake8~=4.0.1", "mypy~=0.960"]

[project.scripts]
libvirt-aws = "libvirt_aws.main:main"

[build-system]
requires = ["versioningit ~= 1.0", "flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "libvirt_aws"

[tool.black]
line-length = 79
target-version = ["py39"]

[tool.mypy]
files = "libvirt_aws"
python_version = "3.9"
follow_imports = "normal"
ignore_missing_imports = true
show_error_codes = true
strict = true

[tool.versioningit]

[tool.versioningit.write]
file = "libvirt_aws/_version.py"
